Introduce ActiveNotificationModel data model
This class serves as an immutable snapshot of ListEntry, so that the
recommended architecture stack doesn't have to worry about the
notification pipeline mutating data out from under it.

This CL also renames and unifies some of the various
Notification-related repositories and interactors; "Active" is used to
identify the outputs of the notification pipeline, which internally
tracks "inactive" notifications (ones that are filtered from view).

Flag: NOTIFICATION_ICON_CONTAINER_REFACTOR
Bug: 290787599
Bug: 278765923
Test: atest SystemUITests
Change-Id: I246b5a5e3a744caad4dcc252c96166e91fe2cc48