Separate the tasks on the main sync and test worker thread
We'll need two threads here since
enableWindowMagnificationWithOffset_expectedValues will require some
time for updating the layout on the screen. postDelayed does not help
here since the test will end directly instead of waiting for the delayed
task to complete.
Therefore, we would like to return to previous `runOnMainSync` way to
only put the required tasks on the main thread and wait on the test
worker thread to verify the result.

Bug: 303132156
Bug: 305732249
Flag: NA
Test: atest SystemUITests:WindowMagnificationAnimationControllerTest
Change-Id: I661a1a6722e9686cdae62a0c8dcb080c46d7e868