refactor(force invert): use enum/intdef for force dark type in setForceDark() functions
This allows the RenderNode to differentiate between normal force dark
and force force-dark (force invert color). In the next CL we will treat
force invert slightly differently, not allowing devs to opt-out of force
dark in individual Views/RenderNodes.

This also sets us up to specify a "force light" mode for force invert,
if we ever need it.

Bug: 282821643
Test: atest ViewRootImplTest
Change-Id: I8cc0fe099cccabdd09ea072aca1e70527e91e1a8