Don't let Keyguard play un/occlude when not visible
When SysUI marks the device as unlocked without going through
keyguardGoingAway, the system server may still be in the process
of creating a transition and that transition will include
information about whether an occluding/unoccluding app is coming
in. Drop the transition from KeyguardTransitionHandler so that it
doesn't get confused and start showing the bouncer again.

It will still get to commit the state after all transitions play.

Change-Id: Iabce80a9f08829750c90acaafb3f3f3fb5f225dc
Test: atest CtsWindowManagerDeviceKeyguard:KeyguardTests
Fix: 303056717