Add "allow channels" to ZenPolicy, NM.Policy, ZenModeConfig
The primary functionality is in ZenPolicy, where the enum for types of channels that may be allowed (unset, priority, none) is defined.

The setting has to be included in NM.Policy and ZenModeConfig as well to allow it to apply for the consolidated policy and the manual rule.

Updates ZenModeFiltering to respect this value when determining whether to intercept a given notification, and PreferencesHelper and ZenModeConfig to accurately report whether any channels are actually potentially bypassing DND given the existence of priority channels and the value of whether channels are permitted.

Bug: 308672314
Bug: 300477976
Test: atest ZenPolicyTest, ZenModeConfigTest, ZenModeDiffTest, ZenModeHelperTest, ZenModeFilteringTest, PreferencesHelperTest
Test: (cts) NotificationManagerZenTest, ZenPolicyTest

Change-Id: Ia774852770df08fb5493be429819b2be39e201de