Remove appLayerReplacesLauncher test from transfer splash screen test.
The splash screen layer can overlay with app window, so actually the
"ignoreSplashscreen" parameter should be true, and better to set
isOptional to false in isSplashScreenVisibleFor. But while using
SurfaceView to draw splash screen icon, not every layer need to be
visible, so it won't pass isSplashScreenVisibleFor assert.
Overall, seems appWindowAfterSplash can cover the original purpose
of this test, so remove appLayerReplacesLauncher for transfer splash
screen test.

Bug: 307366441
Test: atest OpenTransferSplashscreenAppFromLauncherTransition
Change-Id: If4820f8fa7cdfaf2b7a6f7b4eb24b3a818acc101