Migrate A11yService warning resources from Settings to frameworks/base.
This dialog is unnecessarily duplicated inside frameworks/base and
Settings, so this topic is consolidating the dialog to remove the
Settings version. However, the Settings version is more feature-full
so this change copies that version's resources to frameworks/base.

These resources are all not APIs so no Trunk Stable feature flag.

Test: Used in followup change & tested there
Bug: 303511250
Change-Id: I19e660baf47a8c25936b92df22b8ea536af8ddfc