Smooth discontinuous height changes to KeyguardStatusViewController
In some cases, KeyguardStatusViewController's height changes suddenly
when smartspace enters or exits. While smartspace animates it's entries
and exits smoothly, the height and subsequent layout changes are not
included in this. This solves the effects of that problem by smoothing
the height reported to NPVC. This fixes a visual glitch where NSSL can
jump during a doze transition if smartspace enters or exits at the
same time.

Flag: NONE
Bug: 301491166
Test: Manually checked several animation scenarios on felix and shiba
Change-Id: I9a863ef5b740c613ea016eb309034505eac0b518