Always defer remove snapshot starting window if it contains IME
The ActivityRecord#mStartingData would already be removed when calling
to removeStartingWindow, so check StartingData#hasImeSurface is useless
Instead, we should pass that information when decideing to defer remove
snapshot starting window for IME.
Also since we are moving a task to front, should not check the
softInputMode attribute since the status may already changed from
previous launch.

Bug: 310467880
Test: launch app which softInputMode set to stateAlwaysHidden, invoke
IME window, move the task to back then move it to front.
Verify the snapshot starting window will wait for IME show signal.

Change-Id: If829016d5a1a486b3a424fc6add32ceb5d2e3f05