Create old or new provider in Bubbles and CentralSurfacesImpl tests
Right now, these tests create a NotificationInterruptStateProvider
manually, and therefore they have to force disable the flag for the
refactored version of that code.

As we roll out the refactor, we want test coverage of the new code, so
change these tests to create the old or new code based on the flag
value.

Bug: 312476335
Test: atest BubblesTest CentralSurfacesImplTest
Flag: ACONFIG com.android.systemui.visual_interruptions_refactor STAGING
Change-Id: I61900e55fc8c4c6d4f08ffaae11ef98fcf4c50d7