Support more profile tabs
This is specifically for the private profile.

Currently, the ProfileSelectFragment assumes that two tabs need to be
shown. Extending that to allow variable number of tabs depending on the
client's needs.
It still assumes that if client doesn't specify the number of tabs then
two tabs need to be shown - which matches the current impl and allows
the current clients to keep functioning w/o any changes needed.

Bug: 309402121
Test: manual
Change-Id: Ia544ffdd7e0799bcd2e2a9f8f95cc283bb718d91