Use the package added/removed broadcasts for UidState
It was recently changed to use the PackageManagerInternal#getPackageList
callback to manage which packages exist. This API isn't sufficient since
it doesn't provide information about the user so it's impossible to know
which user a package was removed for. We're reverting to the old
behavior that uses the added/removed broadcast which has the UID so we
can properly manage which objects need to be removed.

Test: AppOpsMultiuserTest
Bug: 314689371
Change-Id: Ibd745ce731b4fb433e044956acf5c5427545fbdf