Fix race condition between PromptIconViewBinder and BiometricSizeBinder
Sets BiometricPrompt view invisible until all resizing is complete. Adds
new MutableStateFlow updated on iconView onCompositionLoadedListener,
and combines this new flow with the BiometricSizeBinder viewModel.size
collect call, to force the resizing logic to wait for the iconAsset to
update first to account for it in the prompt size.

Flag: NONE
Fixes: 313605043
Test: (manual) manually verified prompt appears only once and in the correct size, no resizing visible to the user
Change-Id: I6b25acbf779c7f2b80b4f2fe8693c6021dbf2903