Tweak HUN intro/outro animations
Sync the height and alpha animation with the same interpolator, and
adjust the range of the alpha interpolation to avoid collision with the
height clipping and the content.
Add a translateY animation to move in from (and move out to) the start
of the screen.

Fixes: b/243302608
Test: atest StackStateAnimatorTest StackScrollAlgorithmTest
Test: post HUNs and observe the appear/disappear animations
Flag: LEGACY IMPROVED_HUN_ANIMATIONS DEVELOPMENT
Change-Id: I31a42f1e0af0c0ca5df5b8a53b260aaa79807d32