Remove setAutomaticZenRuleState(boolean fromUser) overload
This is a leftover from the first approach, where fromUser was always false and Condition.source not directly checked in NMS. With both changed, it doesn't make sense to supply the exact same information in two parameters (and validate they match!); we can rely on Condition.source only.

(Also resolve a "follow-up" comment from the previous CL about TextUtils.formatSimple).

Bug: 308670715
Test: atest NotificationManagerZenTest
Change-Id: I0e49427636c6735748d9157262c8dd8c76fc7a56