Introducing the SeekableSliderHapticPlugin for volume slider haptics.
The plugin is connected to the VolumeDialogImpl and gets applied to all
the vertical volume sliders. Each plugin allows for dynamic haptic
feedback to be delivered from the interactions with a slider. The
VolumeDialogController also defines a callback that lets the sliders
know when the volume changed due to a physical button.

Test: atest SystemUiRoboTests:SeekableSliderHapticPluginTest
Test: atest SystemUITests:VolumeDialogImplTest
Flag: ACONFIG com.android.systemui.haptic_volume_slider DEVELOPMENT
Bug: 316953430

Change-Id: I476f239bfed3b0e1c8c83c64592974ab0f6e8411