add API for getRegisteredPhoneAccounts
currently there is no way for VoIP applications to fetch their
registered PhoneAccounts with Telecom. TelecomManager#getPhoneAccount
is gated by the READ_PHONE_STATE permission which is a highly privileged
permission that VoIP apps will not be granted. This API will not only allow
VoIP apps to get all registered PhoneAccounts with Telecom but Managed apps
(like system Dialers) can use it.  This is helpful to know what calling
capabilities are currently registered and know the state of Telecom.

Bug: 307609763
Test: CTS coverage
Change-Id: I73ace8cd766369f877c7fc0ef985f6a0c575e097