[flexiglass] Add nested scrolling to NSSL and notif scrim in Flexiglass
Allows NSSL to scroll inside the notification scrim when its contents exceed the height of the scrim, and the scrim to expand to cover the entire screen (excluding the Shade Header) as the NSSL scrolls.

- The scrim is able to expand even when NSSL itself isn't scrollable (i.e. when its contents exceed the height of the "collapsed" scrim but not the height of the "expanded" scrim).
- Single-finger swipes to expand notifications instead scroll the NSSL/scrim when the scrim is expanded.
- While the scrim is expanded, if the NSSL's contents become shorter than the default collapsed scrim height (usually due to notifications being dismissed or collapsed), the scrim will collapse.

Touch handling on notifications that have just been received is fixed in this CL as well.

Bug: 296118689
Test: TBD
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Change-Id: If2620821b7493dfa37e478eb3be8c1de34ca541e