Rename perfetto proto package imported in the framework
perfetto.protos is not an authorized package to be included in the bootclsspath. And because bootclasspath always takes precedence during class resolving, we don't want that when the perfetto library is imported somewhere else (e.g. in an app) that the class name be resolved to class in bootclasspath instead.

Bug: 309630341

Test: presubmit checks
Change-Id: I0c19bc503905cbb33e34240bd6ba970217ee3f29