[2/n] Add configurable font_scale
Adds the `def_device_font_scale` default property to set the initial value
for the font_scale property.

This is used when the device is configured for the first time or
when a new user is created.

The value is read from the def_font_scale resource and stored
in Settings using the Settings.System.DEFAULT_DEVICE_FONT_SCALE key.

Bug:316160749
Test: adb shell settings get system font_scale
Test: atest FrameworksCoreTests:FontScaleConverterFactoryTest

Change-Id: I21e0f04605c43f238440fed535c0c0365d700cb9