Sticky keys UI: implementing repository and view model
Adding StickyKeysRepository which registers listener in InputManager and passes data to flow listeners.
StickyKeyIndicatorViewModel is the only layer above (which is likely to change soon when more input data is needed), domain layer seemed a bit of an overkill for now.

Classes are not used from anywhere yet so code is not active or flagged.

Test: StickyKeysIndicatorViewModelTest
Bug: 313855932
Flag: None

Change-Id: I5493d866947d8553f8b40f7fb26324eafaf312a6