Handle adding a bubbling during drag
When a bubble is added, we reorder bubbles which removes them from the view hierarchy and cancelling all touch events. If a bubble is added while dragging, the gesture is abruptly cancelled and the stack remains in a bad state.

This change ensures that the reordering happens at the end of the gesture.

Fixes: 317847303
Test: Manual
      - Add bubbles
      - Expand the stack and start dragging a bubble
      - While dragging trigger a new bubble notification
      - Observe that new bubble is added and dragged bubble still follows the gesture
      - Release dragged bubble either to dismiss it or to return it to the stack
Change-Id: I30072aeb506edb34a023f44728f838b5b3de184e